--- example/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "compressjs-example",
-  "version": "1.0.0",
-  "description": "Example compress nodejs library",
-  "main": "dist/compress-example.js",
-  "module": "dist/compress-example.module.js",
-  "unpkg": "dist/compress-example.umd.js",
-  "source": "index.js",
-  "files": [
-    "dist/**/*"
-  ],
-  "scripts": {
-    "prepare-commitlint": "npx husky install && chmod ug+x ./.husky/*",
-    "build": "microbundle",
-    "dev": "microbundle watch",
-    "prepublishOnly": "npm run build",
-    "prepare": "npm run build"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/HiWay-Media/compress-nodejs.git"
-  },
-  "keywords": [
-    "compressjs",
-    "nodejs",
-    "typescript",
-    "tangram",
-    "compress"
-  ],
-  "author": "allan nava",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/HiWay-Media/compress-nodejs/issues"
-  },
-  "homepage": "https://github.com/HiWay-Media/compress-nodejs#readme",
-  "dependencies": {
-    "compress-nodejs": "^0.5.0",
-    "microbundle": "^0.15.1",
-    "evaporate": "^2.1.4"
-  }
-}
