--- lib/compressjs/index.ts
@@ -1,16 +1,0 @@
-/**
- * 
- * Author: Allan Nava       (allan.nava@hiway.media)
- * Author: Antonio Borgese  (antonio.borgese@hiway.media)
- * -----
- * Last Modified: 
- * Modified By: Allan Nava (allan.nava@hiway.media>)
- * -----
- * Copyright 2023 - 2023 © 
- * 
- * 
- */
-export * from "./constants";
-export * from "./fetch-wrapper";
-export * from "./tangram";
-//
