--- lib/index.ts
@@ -1,14 +1,0 @@
-/**
- * v0.5.04
- * Author: Allan Nava       (allan.nava@hiway.media)
- * Author: Antonio Borgese  (antonio.borgese@hiway.media)
- * -----
- * Last Modified: 
- * Modified By: Allan Nava (allan.nava@hiway.media>)
- * -----
- * Copyright 2023 - 2024 © 
- * 
- * 
- */
-export * from "./compressjs";
-//
