--- tsconfig.json
@@ -1,35 +1,0 @@
-{
-    "include": [
-      "lib/**/*.ts",
-      "lib/**/*.tsx"
-    ],
-    "exclude": [
-      "node_modules",
-      "dist",
-    ],
-    "compilerOptions": {
-      "target": "es5",
-      "module": "esnext",
-      "lib": [
-        "dom.iterable",
-        "dom",
-        "esnext"
-      ],
-      "allowJs": false,
-      "jsx": "react",
-      "declaration": true,
-      "sourceMap": true,
-      "outDir": "dist",
-      "rootDir": "lib",
-      "baseUrl": "lib",
-      "removeComments": true,
-      "strict": false,
-      "moduleResolution": "node",
-      "allowSyntheticDefaultImports": true,
-      "esModuleInterop": true,
-      "resolveJsonModule": true,
-      "skipLibCheck": true,
-      "forceConsistentCasingInFileNames": true
-    }
-  }
-
